Move over dar tests, follow up with new list tests for arbitrary uri
